test/ui_test/requirements.txt
python_version >= '3.9'

Appium-Python-Client>=4.2.0